Update dependencies and added formatting
Formatted f-strings of simulation output to 11 digits (if you know pie past 11 digits, good for you. In that case, you should be able to find out how to change that to however many digits you want). Updated imports of skimage functions to reflect their reorganization of circle to circle_perimeter and morphology.watershed to segmentation.watershed